Add interaction UI to installer application
If you use PC to burn the ISO image to USB stick,
and forget to plug out the USB stick from PC,
then reboot your PC with USB stick plugged.
Celadon ISO image will be installed to the PC
automatically without notification. Thus it will
damage the data in your PC.

So we add UI to ask the user to confirm if continue
or stop flashing celadon images to this device to avoid
data damage.

Tracked-On: OAM-103788
Signed-off-by: Chen, Gang G <gang.g.chen@intel.com>